Revert "Add support for Cordova + hot reloading"
