+++ .eslintrc
@@ -30,8 +30,6 @@
     "node": true
   },
   "globals": {
-    '__CORDOVA__': true,
-    '__BASE_URL__': true,
     expect: true
   },
   "extends": "eslint:recommended",
