--- common/layouts/cordova.html
@@ -1,24 +1,0 @@
-<!doctype html>
-<html>
-<head>
-<title><%= htmlWebpackPlugin.options.title %></title>
-<meta charset="utf-8" />
-<meta name="viewport" content="initial-scale=1, maximum-scale=1, user-scalable=no, width=device-width height=device-height" />
-<!-- Cordova app security policy:
-    * gap: is required only on iOS (when using UIWebView) and is needed for JS->native communication
-    * https://ssl.gstatic.com is required only on Android and is needed for TalkBack to function properly
-    * Disables use of eval() and inline scripts in order to mitigate risk of XSS vulnerabilities. To change this:
-        * Enable inline JS: add 'unsafe-inline' to default-src
-        * Enable eval(): add 'unsafe-eval' to default-src
--->
-<meta http-equiv="Content-Security-Policy" content="default-src * data: gap: https://ssl.gstatic.com; style-src * 'unsafe-inline'; script-src * 'unsafe-inline' 'unsafe-eval'">
-<meta name="apple-mobile-web-app-capable" content="yes">
-<link type="text/css" rel="stylesheet" href="<%= htmlWebpackPlugin.options.stylesheet %>" />
-</head>
-<body>
-<div id="app"></div>
-<script>window.__INITIAL_STATE__ = {};</script>
-<script src="cordova.js"></script>
-<script src="<%= htmlWebpackPlugin.options.bundleJs %>"></script>
-</body>
-</html>
