+++ package.json
@@ -13,35 +13,24 @@
     "npm": "3.x.x"
   },
   "scripts": {
-    "start": "npm run dev:build && npm run dev:start",
-    "dev:start": "npm run dev:start:client & npm run dev:start:server",
-    "dev:start:cordova": "npm run cordova:dev:build:hot; npm run dev:start:server & npm run dev:start:cordova:client",
-    "dev:start:cordova:client": "NODE_ENV=development node webpack/config.dev.cordova.hot",
-    "dev:start:server": "NODE_ENV=development $(npm bin)/nodemon server",
-    "dev:start:client": "NODE_ENV=development node webpack/config.dev.hot",
-    "dev:build": "NODE_ENV=development $(npm bin)/webpack --progress --colors --display-error-details --config webpack/config.dev.js",
-    "prod:build": "NODE_ENV=production $(npm bin)/webpack -p --optimize-dedupe --config webpack/config.prod.js",
-    "prod:start": "NODE_ENV=production pm2 start server --name='callai'",
-    "cordova:dev:build:hot": "NODE_ENV=development npm run cordova:clean; HOT=true $(npm bin)/webpack --progress --colors --display-error-details --config webpack/config.dev.cordova.js; npm run cordova:prepare",
-    "cordova:dev:build": "NODE_ENV=development npm run cordova:clean; $(npm bin)/webpack --progress --colors --display-error-details --config webpack/config.dev.cordova.js; npm run cordova:prepare",
-    "cordova:prod:build": "NODE_ENV=production npm run cordova:clean; NODE_ENV=production $(npm bin)/webpack -p --optimize-dedupe --config webpack/config.prod.cordova.js; npm run cordova:prepare;",
-    "cordova:dist": "npm run cordova:dist:android && npm run cordova:dist:ios",
-    "cordova:dist:android": "node ./scripts/cordova --platform android --config release --distribute",
-    "crodova:dist:ios": "node ./scripts/cordova --platform ios --config release --distribute",
-    "cordova:clean": "rm -f $(find cordova/www/* 2>/dev/null -type f | grep -v 'gitkeep')",
-    "cordova:prepare": "cd cordova && cordova prepare",
+    "start": "npm run build:development && npm run start:development",
+    "start:production": "NODE_ENV=production node server",
+    "start:development": "NODE_ENV=development npm run start:development:client & npm run start:development:server",
+    "start:development:client": "node webpack/config.development.hot",
+    "start:development:server": "$(npm bin)/nodemon server",
+    "build:development": "$(npm bin)/webpack --progress --colors --display-error-details --config webpack/config.development.js",
+    "build:production": "NODE_ENV=production $(npm bin)/webpack -p --optimize-dedupe --config webpack/config.production.js",
     "test": "$(npm bin)/mocha --recursive --compilers js:babel-register",
     "test:all": "$(npm bin)/mocha --recursive --compilers js:babel-register $(find . -path ./node_modules -prune -o -name '*.test.js' -print)",
     "test:watch": "npm test -- --watch",
     "test:all:watch": "$(npm bin)/mocha --recursive --compilers js:babel-register $(find . -path ./node_modules -prune -o -name '*.test.js' -print) --watch",
     "lint": "$(npm bin)/eslint --ignore-pattern **/node_modules/* --ext .js,.jsx .",
-    "postinstall": "if [ \"$NODE_ENV\" = \"production\" ]; then npm run prod:build; fi"
+    "postinstall": "if [ \"$NODE_ENV\" = \"production\" ]; then npm run build:production; fi"
   },
   "devDependencies": {
     "babel-eslint": "6.0.4",
     "chai": "3.5.0",
     "chai-immutable": "1.5.4",
-    "cordova-lib": "6.2.0",
     "eslint": "2.12.0",
     "eslint-plugin-react": "5.1.1",
     "mocha": "2.5.3",
