+++ webpack/config.base.js
@@ -2,7 +2,7 @@ const path = require('path');
 const webpack = require('webpack');
 const isomorphicConfig = require('./config.isomorphic');
 const IsomorphicPlugin = require('webpack-isomorphic-tools/plugin');
-const host = require('./host')();
+const host = require('./host.devserver');
 
 const isDev = process.env.NODE_ENV === 'development';
 const isomorphicPlugin = new IsomorphicPlugin(isomorphicConfig).development(isDev);
@@ -47,8 +47,7 @@ module.exports = {
     new webpack.DefinePlugin({
       'process.env': {
         'NODE_ENV': JSON.stringify(process.env.NODE_ENV)
-      },
-      '__CORDOVA__': false
+      }
     })
   ]
 };
