+++ webpack/config.development.hot.js
@@ -2,23 +2,44 @@
 const webpack = require('webpack');
 const WebpackDevServer = require('webpack-dev-server');
 const baseConfig = require('./config.base');
-const host = require('./host')({ OUTPUT_PATH: 'cordova/www/' });
-const hot = require('./hot');
+const host = require('./host.devserver');
 
-const entry = hot.entry(host.HOST_URL);
-const plugins = [
-  ...hot.plugins,
-  new webpack.DefinePlugin({
-    'process.env': { 'NODE_ENV': JSON.stringify('development') },
-    '__CORDOVA__': true
-  })
+const entry = [
+  'webpack-dev-server/client?' + host.URL,
+  'webpack/hot/only-dev-server'
 ];
-const loaders = hot.loaders;
+
+const plugins = [
+  new webpack.optimize.OccurenceOrderPlugin(),
+  new webpack.HotModuleReplacementPlugin(),
+  new webpack.NoErrorsPlugin()
+];
+
+const loaders = [{
+  test: /\.scss$/,
+  loader: 'style!css!sass'
+}, {
+  test: /\.jsx$|\.js$/,
+  loader: 'babel',
+  exclude: /node_modules/,
+  // use react-transform to hot reload modules when hot is specified
+  query: {
+    plugins: [
+      ['react-transform', {
+        transforms: [{
+          transform: 'react-transform-hmr',
+          imports: ['react'],
+          locals: ['module']
+        }, {
+          transform: 'react-transform-catch-errors',
+          imports: ['react', 'redbox-react']
+        }]
+      }]
+    ]
+  }
+}];
 
 const config = Object.assign({}, baseConfig, {
-  output: Object.assign({}, baseConfig.output, {
-    publicPath: host.ASSET_HOST
-  }),
   eslint: { configFile: './.eslintrc' },
   devServerPort: host.PORT,
   devtool: 'eval',
@@ -38,7 +59,7 @@ const config = Object.assign({}, baseConfig, {
   })
 });
 
-console.info('Firing up Cordova Webpack dev server...');
+console.info('Firing up Wepback dev server...');
 
 new WebpackDevServer(webpack(config), {
   publicPath: config.output.publicPath,
@@ -58,6 +79,6 @@ new WebpackDevServer(webpack(config), {
   if (err) {
     console.error(err);
   } else {
-    console.info('🚧 Cordova-Webpack client dev-server listening on ' + host.HOST_URL + ' with publicPath:' + config.output.publicPath);
+    console.info('🚧 Webpack client dev-server listening on ' + host.URL + ' with publicPath:' + config.output.publicPath);
   }
 });
